+++ src/core/horizon/filesystem/memory_file.hpp
@@ -7,15 +7,15 @@ namespace hydra::horizon::filesystem {
 class MemoryFile : public IFile {
   public:
     MemoryFile(const std::vector<u8>& data_) : data{std::move(data_)} {}
-    MemoryFile(usize size) : data(size) {}
+    MemoryFile(u64 size) : data(size) {}
 
-    void Resize(usize new_size) override { data.resize(new_size); }
+    void Resize(u64 new_size) override { data.resize(new_size); }
 
     io::IStream* Open([[maybe_unused]] FileOpenFlags flags) override {
         return new io::MemoryStream(data);
     }
 
-    usize GetSize() const override { return data.size(); }
+    u64 GetSize() const override { return data.size(); }
 
   private:
     std::vector<u8> data;
