+++ src/core/hw/tegra_x1/cpu/const.hpp
@@ -2,7 +2,7 @@
 
 namespace hydra::hw::tegra_x1::cpu {
 
-constexpr usize GUEST_PAGE_SIZE = 0x1000;
+constexpr u64 GUEST_PAGE_SIZE = 0x1000;
 
 constexpr u32 MAX_STACK_TRACE_DEPTH = 32;
