+++ src/core/hw/tegra_x1/cpu/dynarmic/mmu.hpp
@@ -5,7 +5,7 @@
 
 namespace hydra::hw::tegra_x1::cpu::dynarmic {
 
-constexpr usize PAGE_COUNT =
+constexpr u64 PAGE_COUNT =
     horizon::kernel::ADDRESS_SPACE.GetEnd() / GUEST_PAGE_SIZE;
 
 class Mmu : public IMmu {
@@ -19,7 +19,7 @@ class Mmu : public IMmu {
     void Protect(Range<vaddr_t> range,
                  horizon::kernel::MemoryPermission perm) override;
 
-    void ResizeHeap(IMemory* heap_mem, vaddr_t va, usize size) override;
+    void ResizeHeap(IMemory* heap_mem, vaddr_t va, u64 size) override;
 
     uptr UnmapAddr(vaddr_t va) const override;
     MemoryRegion QueryRegion(vaddr_t va) const override;
